fix(server): review round 8 — incarnation-fenced revocation, freshness-gated listings, completion-only retries
- IssuedJti records minted_at; departure revocation partitions the
  bucket at the departure-observation instant so a concurrently minted
  replacement token survives un-revoked, suppresses the pending eject,
  and keeps last_minted_at for the supersession fence
- CallEntry.room_sid_observed_at + listing_started_at gate
  rotate_room_incarnation_from_listing so a stale ListRooms snapshot
  cannot roll the sid fence back onto a webhook-restored newer room
- ListedRoom.name is the typed ListedRoomName enum (Waddle(CallId) /
  Foreign) classified once at admin deserialization
- new completion-only TeardownTarget::CallThreadEndRetry replaces the
  destructive MujiPresenceClear replay for call-thread completion
  retries from the MUC presence path
- MujiRoomSweep rows without a room SID dead-letter as failed
  (CallTeardownLastError::MissingRoomSid) instead of silently
  completing; new IntentExecution::Permanent drain arm
- infrastructure-transient retry reasons (LiveKit executor/admin
  unavailable) keep requeueing at the capped backoff instead of
  dead-lettering at MAX_ATTEMPTS; new coverage for both behaviors
- the 1:1 terminate rate limiter is charged only on authorized
  mutating teardowns, so bogus unknown-session terminates cannot
  starve a legitimate hangup from another resource
- webhook delivery retention pruning is amortized to one observe per
  minute via a store-held CAS slot
- call_teardown_outbox/tests.rs split into concern-scoped modules
