+++ server/crates/waddle-server/src/call_teardown_outbox/schema.rs
@@ -13,7 +13,7 @@ pub(super) async fn initialize(db: &Database) -> Result<(), CallTeardownOutboxEr
                     call_id TEXT NOT NULL, \
                     identity TEXT NULL CHECK (identity IS NULL OR identity <> ''), \
                     room_jid TEXT NULL CHECK (room_jid IS NULL OR room_jid <> ''), \
-                    action TEXT NOT NULL CHECK (action IN ('remove_participant','delete_room','muji_presence_clear','muji_room_sweep')), \
+                    action TEXT NOT NULL CHECK (action IN ('remove_participant','delete_room','muji_presence_clear','muji_room_sweep','call_thread_end_retry')), \
                     generation INTEGER NULL CHECK (generation IS NULL OR generation > 0), \
                     room_sid TEXT NULL, \
                     participant_sid TEXT NULL, \
@@ -30,7 +30,8 @@ pub(super) async fn initialize(db: &Database) -> Result<(), CallTeardownOutboxEr
                         (action = 'remove_participant' AND identity IS NOT NULL AND room_jid IS NULL) \
                         OR (action = 'delete_room' AND identity IS NULL AND room_jid IS NULL AND participant_sid IS NULL) \
                         OR (action = 'muji_presence_clear' AND identity IS NOT NULL AND room_jid IS NOT NULL) \
-                        OR (action = 'muji_room_sweep' AND identity IS NULL AND room_jid IS NOT NULL AND room_sid IS NOT NULL AND participant_sid IS NULL)\
+                        OR (action = 'muji_room_sweep' AND identity IS NULL AND room_jid IS NOT NULL AND room_sid IS NOT NULL AND participant_sid IS NULL) \
+                        OR (action = 'call_thread_end_retry' AND identity IS NULL AND room_jid IS NOT NULL AND participant_sid IS NULL)\
                     )\
                 )"
             ),
